requirements.txt
accelerate==1.7.0
bitsandbytes
fastchat
huggingface_hub
loralib
numpy==1.26.4
torch==2.3.0
tqdm
transformers==4.50.0
deepspeed==0.13.5
wheel
wandb
flash-attn==2.7.4.post1
peft==0.13.2
datasets
scipy
scikit-learn
# PDF processing dependencies
PyPDF2
pdfplumber
pymupdf
pillow